refactor: move Search button to new component
